+++ lib/commands/commands.dart
@@ -1,0 +1,3 @@
+export 'create_command.dart';
+export 'help_command.dart';
+export 'connect_command.dart';