+++ CHANGELOG.md
@@ -1,4 +1,5 @@
-## 1.0.0
-- Initial version.
+## 0.0.1
+- added possibility to create a Flutter project template.
+- added the ability to connect a remote Github repository.
 
-## 1.0.1
+